chore: move all hooks to root-layer
