+++ components/index.ts
@@ -188,8 +188,8 @@ export { default as useKeyboard, KeyMod, KeyCode } from './use-keyboard'
 export type { KeyboardOptions, UseKeyboardHandler } from './use-keyboard'
 
 export { default as useInput } from './use-input'
-export { default as useModal } from './modal/use-modal'
-export { default as useTabs } from './tabs/use-tabs'
+export { default as useModal } from './use-modal'
+export { default as useTabs } from './use-tabs'
 export { default as useClickAway } from './use-click-away'
 export { default as useCurrentState } from './use-current-state'
 export { default as CssBaseline } from './css-baseline'
