+++ components/use-modal/index.ts
@@ -1,0 +1,4 @@
+import useModal from './use-modal'
+
+export type { ModalHooksBindings } from './use-modal'
+export default useModal
