+++ components/use-tabs/index.ts
@@ -1,0 +1,3 @@
+import useTabs from './use-tabs'
+
+export default useTabs
